Refactor feeder device classes by consolidating shared functionality into a base class (`Feeder`) and removing redundant code.
