+++ custom_components/petlibro/devices/feeders/__init__.py
@@ -1,1 +1,1 @@
-from .. import Device
+from .feeder import Feeder
