+++ custom_components/petlibro/devices/feeders/polar_wet_food_feeder.py
@@ -1,35 +1,30 @@
 import aiohttp
-
-from ...api import make_api_call
-from aiohttp import ClientSession, ClientError
 import asyncio
 from datetime import datetime
-from ...exceptions import PetLibroAPIError
-from ..device import Device
+
 from typing import cast
 from logging import getLogger
+from ...exceptions import PetLibroAPIError
+from .feeder import Feeder
 
 _LOGGER = getLogger(__name__)
 
-class PolarWetFoodFeeder(Device):
+
+class PolarWetFoodFeeder(Feeder):
     async def refresh(self):
         """Refresh the device data from the API."""
         try:
-            await super().refresh()  # This calls the refresh method in GranaryFeeder (which also inherits from Device)
-    
+            await super().refresh()
+
             # Fetch specific data for this device
             grain_status = await self.api.device_grain_status(self.serial)
-            real_info = await self.api.device_real_info(self.serial)
-            attribute_settings = await self.api.device_attribute_settings(self.serial)
             get_upgrade = await self.api.get_device_upgrade(self.serial)
             wet_feeding_plan = await self.api.device_wet_feeding_plan(self.serial)
             get_feeding_plan_today = await self.api.device_feeding_plan_today_new(self.serial)
-    
+
             # Update internal data with fetched API data
             self.update_data({
                 "grainStatus": grain_status or {},
-                "realInfo": real_info or {},
-                "getAttributeSetting": attribute_settings or {},
                 "getUpgrade": get_upgrade or {},
                 "wetFeedingPlan": wet_feeding_plan or {},
                 "getfeedingplantoday": get_feeding_plan_today or {}
@@ -37,34 +32,18 @@ class PolarWetFoodFeeder(Device):
         except PetLibroAPIError as err:
             _LOGGER.error(f"Error refreshing data for PolarWetFoodFeeder: {err}")
 
-    @property
-    def available(self) -> bool:
-        _LOGGER.debug(f"Device {self.device.name} availability: {self.device.online}")
-        return self.device.online if hasattr(self.device, 'online') else True
+    # ------------------------------------------------------------------
+    # Overrides: Polar reads these from different data paths
+    # ------------------------------------------------------------------
 
     @property
     def battery_state(self) -> str:
-        return cast(str, self._data.get("batteryState", "unknown"))  # Battery status is low or unknown
-    
-    @property
-    def battery_display_type(self) -> float:
-        """Get the battery percentage state."""
-        try:
-            value = str(self._data.get("realInfo", {}).get("batteryDisplayType", "percentage"))
-            # Attempt to convert the value to a float
-            return cast(float, float(value))
-        except (TypeError, ValueError):
-            # Handle the case where the value is None or not a valid float
-            return 0.0
+        return cast(str, self._data.get("batteryState", "unknown"))
 
     @property
     def device_sn(self) -> str:
         """Returns the serial number of the device."""
         return self._data.get("deviceSn", "unknown")
-
-    @property
-    def door_blocked(self) -> bool:
-        return bool(self._data.get("realInfo", {}).get("barnDoorError", False))
 
     @property
     def electric_quantity(self) -> float:
@@ -73,18 +52,31 @@ class PolarWetFoodFeeder(Device):
         return quantity if isinstance(quantity, (float, int)) else 0
 
     @property
-    def feeding_plan_state(self) -> bool:
-        """Return the state of the feeding plan."""
-        return bool(self._data.get("enableFeedingPlan", False))
-
-    @property
     def food_low(self) -> bool:
-        return not bool(self._data.get("surplusGrain", True))  # Surplus grain available
+        return not bool(self._data.get("surplusGrain", True))
 
     @property
     def mac_address(self) -> str:
         """Returns the MAC address of the device."""
         return self._data.get("mac", "unknown")
+
+    @property
+    def online(self) -> bool:
+        """Returns the online status of the device."""
+        return self._data.get("online", False)
+
+    @property
+    def wifi_rssi(self) -> int:
+        wifi_rssi = self._data.get("wifiRssi")
+        return wifi_rssi if isinstance(wifi_rssi, int) else -100
+
+    # ------------------------------------------------------------------
+    # Polar-specific properties
+    # ------------------------------------------------------------------
+
+    @property
+    def door_blocked(self) -> bool:
+        return bool(self._data.get("realInfo", {}).get("barnDoorError", False))
 
     @property
     def next_feeding_day(self) -> str:
@@ -98,9 +90,8 @@ class PolarWetFoodFeeder(Device):
         if raw_time == "unknown":
             return raw_time
         try:
-            # Convert 24-hour time to 12-hour format with AM/PM
             time_obj = datetime.strptime(raw_time, "%H:%M")
-            return time_obj.strftime("%I:%M %p")  # "08:00 AM" or "11:00 PM"
+            return time_obj.strftime("%I:%M %p")
         except ValueError:
             return "Invalid time"
 
@@ -111,26 +102,20 @@ class PolarWetFoodFeeder(Device):
         if raw_time == "unknown":
             return raw_time
         try:
-            # Convert 24-hour time to 12-hour format with AM/PM
             time_obj = datetime.strptime(raw_time, "%H:%M")
-            return time_obj.strftime("%I:%M %p")  # "08:00 AM" or "11:00 PM"
+            return time_obj.strftime("%I:%M %p")
         except ValueError:
             return "Invalid time"
-        
+
     @property
     def manual_feed_id(self) -> int:
         """Returns the manual feed ID."""
         return self._data.get("wetFeedingPlan", {}).get("manualFeedId", None)
-        
+
     @property
     def manual_feed_now(self) -> bool:
         """Returns whether the feeder is set to feed now or not."""
         return self.manual_feed_id is not None
-
-    @property
-    def online(self) -> bool:
-        """Returns the online status of the device."""
-        return self._data.get("online", False)
 
     @property
     def online_list(self) -> list:
@@ -147,37 +132,11 @@ class PolarWetFoodFeeder(Device):
         """Returns the current temperature in Fahrenheit, rounded to 1 decimal place."""
         celsius = self._data.get("realInfo", {}).get("temperature", 0.0)
         fahrenheit = celsius * 9 / 5 + 32
-        return round(fahrenheit, 1)  # Round to 1 decimal place
+        return round(fahrenheit, 1)
 
-    @property
-    def unit_type(self) -> int:
-        return self._data.get("realInfo", {}).get("unitType", 1)
-
-    @property
-    def whether_in_sleep_mode(self) -> bool:
-        return bool(self._data.get("getAttributeSetting", {}).get("enableSleepMode", False))
-
-    @property
-    def enable_low_battery_notice(self) -> bool:
-        return bool(self._data.get("realInfo", {}).get("enableLowBatteryNotice", False))
-    
-    @property
-    def wifi_rssi(self) -> int:
-        wifi_rssi = self._data.get("wifiRssi")
-        return wifi_rssi if isinstance(wifi_rssi, int) else -100  # WiFi signal strength
-
-    @property
-    def wifi_ssid(self) -> str:
-        return self._data.get("realInfo", {}).get("wifiSsid", "unknown")
-
-    @property
-    def feeding_plan_today_data(self) -> str:
-        return self._data.get("getfeedingplantoday", {})
-
-    @property
-    def light_switch(self) -> bool:
-        """Check if the light is enabled."""
-        return bool(self._data.get("realInfo", {}).get("lightSwitch", False))
+    # ------------------------------------------------------------------
+    # Polar-specific methods
+    # ------------------------------------------------------------------
 
     async def set_manual_feed_now(self, start: bool, plate: int) -> None:
         plate = plate if plate is not None else self.plate_position
@@ -188,12 +147,12 @@ class PolarWetFoodFeeder(Device):
             else:
                 _LOGGER.debug(f"Triggering stop feed now for {self.serial}")
                 await self.api.set_stop_feed_now(self.serial, self.manual_feed_id)
-            
-            await self.refresh()  # Refresh the state after the action
+
+            await self.refresh()
         except aiohttp.ClientError as err:
             _LOGGER.error(f"Failed to trigger manual feed now for {self.serial} with plate no.{plate}: {err}")
             raise PetLibroAPIError(f"Error triggering manual feed now: {err}")
-    
+
     async def set_plate_position(self, value: str | int) -> None:
         """Rotate bowl to requested plate (1-3)"""
         try:
@@ -201,10 +160,8 @@ class PolarWetFoodFeeder(Device):
         except (TypeError, ValueError):
             raise PetLibroAPIError(f"Invalid plate value: {value!r}")
         if target not in (1, 2, 3):
-            # Raise an error if plate count somehow became less than 1 or more than 3.
             raise PetLibroAPIError(f"Plate must be 1, 2, or 3, got {target}")
 
-        # Ensure we know current position
         if not self.plate_position:
             await self.refresh()
         curr = self.plate_position or 1
@@ -212,14 +169,12 @@ class PolarWetFoodFeeder(Device):
         steps = (target - curr) % 3
         _LOGGER.debug("Rotate-to-plate: curr=%s target=%s steps=%s for %s", curr, target, steps, self.serial)
 
-        # didnt test other cooldowns, may be able reduce.
         ROTATE_COOLDOWN = 0.6
         for _ in range(steps):
             await self.api.set_rotate_food_bowl(self.serial)
             await asyncio.sleep(ROTATE_COOLDOWN)
             await self.refresh()
 
-        # Final refresh so sensor/current_option show the target
         await self.refresh()
 
     async def rotate_food_bowl(self) -> None:
@@ -227,7 +182,7 @@ class PolarWetFoodFeeder(Device):
 
         try:
             await self.api.set_rotate_food_bowl(self.serial)
-            await self.refresh()  # Refresh the state after the action
+            await self.refresh()
         except aiohttp.ClientError as err:
             _LOGGER.error(f"Failed to trigger rotate food bowl for {self.serial}: {err}")
             raise PetLibroAPIError(f"Error triggering rotate food bowl: {err}")
@@ -237,7 +192,7 @@ class PolarWetFoodFeeder(Device):
 
         try:
             await self.api.set_feed_audio(self.serial)
-            await self.refresh()  # Refresh the state after the action
+            await self.refresh()
         except aiohttp.ClientError as err:
             _LOGGER.error(f"Failed to trigger feed audio for {self.serial}: {err}")
             raise PetLibroAPIError(f"Error triggering feed audio: {err}")
@@ -247,7 +202,6 @@ class PolarWetFoodFeeder(Device):
 
         if not self._data.get("wetFeedingPlan"):
             _LOGGER.debug(f"Triggering device data refresh because wet feeding plan data is missing for {self.serial}")
-            # Refresh the state to ensure the wet feeding plan is already fetched
             try:
                 await self.refresh()
             except aiohttp.ClientError as err:
@@ -279,60 +233,7 @@ class PolarWetFoodFeeder(Device):
 
         try:
             await self.api.set_reposition_schedule(self.serial, current_feeding_plan, plan_name)
-            await self.refresh() # Refresh the state after the action
+            await self.refresh()
         except aiohttp.ClientError as err:
             _LOGGER.error(f"Failed to trigger reposition the schedule for {self.serial}: {err}")
             raise PetLibroAPIError(f"Error triggering reposition schedule: {err}")
-
-    # Method for indicator turn on
-    async def set_light_on(self) -> None:
-        _LOGGER.debug(f"Turning on the indicator for {self.serial}")
-        try:
-            await self.api.set_light_on(self.serial)
-            await self.refresh()  # Refresh the state after the action
-        except aiohttp.ClientError as err:
-            _LOGGER.error(f"Failed to turn on the indicator for {self.serial}: {err}")
-            raise PetLibroAPIError(f"Error turning on the indicator: {err}")
-
-    # Method for indicator turn off
-    async def set_light_off(self) -> None:
-        _LOGGER.debug(f"Turning off the indicator for {self.serial}")
-        try:
-            await self.api.set_light_off(self.serial)
-            await self.refresh()  # Refresh the state after the action
-        except aiohttp.ClientError as err:
-            _LOGGER.error(f"Failed to turn off the indicator for {self.serial}: {err}")
-            raise PetLibroAPIError(f"Error turning off the indicator: {err}")
-
-    @property
-    def update_available(self) -> bool:
-        """Return True if an update is available, False otherwise."""
-        return bool(self._data.get("getUpgrade", {}).get("jobItemId"))
-    
-    @property
-    def update_release_notes(self) -> str | None:
-        """Return release notes if available, else None."""
-        upgrade_data = self._data.get("getUpgrade")
-        return upgrade_data.get("upgradeDesc") if upgrade_data else None
-    
-    @property
-    def update_version(self) -> str | None:
-        """Return target version if available, else None."""
-        upgrade_data = self._data.get("getUpgrade")
-        return upgrade_data.get("targetVersion") if upgrade_data else None
-    
-    @property
-    def update_name(self) -> str | None:
-        """Return update job name if available, else None."""
-        upgrade_data = self._data.get("getUpgrade")
-        return upgrade_data.get("jobName") if upgrade_data else None
-    
-    @property
-    def update_progress(self) -> float:
-        """Return update progress as a float, or 0 if not updating."""
-        upgrade_data = self._data.get("getUpgrade")
-        if not upgrade_data:
-            return 0.0
-
-        progress = upgrade_data.get("progress")
-        return float(progress) if progress is not None else 0.0
