+++ custom_components/petlibro/devices/fountains/__init__.py
@@ -1,1 +1,2 @@
 from .. import Device
+from .fountain import Fountain
