+++ custom_components/petlibro/devices/fountains/dockstream_2_smart_cordless_fountain.py
@@ -1,37 +1,29 @@
 import aiohttp
 
-from ...api import make_api_call
-from aiohttp import ClientSession, ClientError
 from ...exceptions import PetLibroAPIError
-from ..device import Device
+from .fountain import Fountain
 from typing import cast
 from logging import getLogger
 
 _LOGGER = getLogger(__name__)
 
-class Dockstream2SmartCordlessFountain(Device):
+class Dockstream2SmartCordlessFountain(Fountain):
     """Represents the Dockstream 2 Smart Cordless Fountain device."""
 
     async def refresh(self):
         """Refresh the device data from the API."""
         try:
-            await super().refresh()  # Call the refresh method from the parent class (Device)
-        
-            # Fetch real info from the API
-            real_info = await self.api.device_real_info(self.serial)
+            await super().refresh()
+
             data_real_info = await self.api.device_data_real_info(self.serial)
-            attribute_settings = await self.api.device_attribute_settings(self.serial)
             get_upgrade = await self.api.get_device_upgrade(self.serial)
             get_work_record = await self.api.get_device_work_record(self.serial)
             get_feeding_plan_today = await self.api.device_feeding_plan_today_new(self.serial)
             get_drink_water = await self.api.get_device_drink_water(self.serial)
 
-            # Update internal data with fetched API data
             self.update_data({
-                "realInfo": real_info or {},
                 "dataRealInfo": data_real_info or {},
                 "getDrinkWater": get_drink_water or {},
-                "getAttributeSetting": attribute_settings or {},
                 "getUpgrade": get_upgrade or {},
                 "getfeedingplantoday": get_feeding_plan_today or {},
                 "workRecord": get_work_record if get_work_record is not None else []
@@ -40,87 +32,12 @@ class Dockstream2SmartCordlessFountain(Device):
         except PetLibroAPIError as err:
             _LOGGER.error(f"Error refreshing data for Dockstream2SmartCordlessFountain: {err}")
 
-    @property
-    def available(self) -> bool:
-        _LOGGER.debug(f"Device {self.device.name} availability: {self.device.online}")
-        return self.device.online if hasattr(self.device, 'online') else True
-
-    @property
-    def device_sn(self) -> str:
-        """Return the device serial number."""
-        return self._data.get("realInfo", {}).get("deviceSn", "unknown")
-
-    @property
-    def wifi_ssid(self) -> str:
-        """Return the Wi-Fi SSID of the device."""
-        return self._data.get("realInfo", {}).get("wifiSsid", "unknown")
-
-    @property
-    def online(self) -> bool:
-        """Return the online status of the fountain."""
-        return bool(self._data.get("realInfo", {}).get("online", False))
-    
-    @property
-    def wifi_rssi(self) -> int:
-        """Get the Wi-Fi signal strength."""
-        wifi_rssi = self._data.get("realInfo", {}).get("wifiRssi")
-        return wifi_rssi if isinstance(wifi_rssi, int) else -100
-    
-    @property
-    def weight(self) -> float:
-        """Get the current weight of the water (in grams)."""
-        weight = self._data.get("realInfo", {}).get("weight")
-        return weight if isinstance(weight, (int, float)) else 0
-    
-    @property
-    def weight_percent(self) -> int | float:
-        """Get the current weight percentage of water."""
-        weight_percent = self._data.get("realInfo", {}).get("weightPercent")
-        return weight_percent if isinstance(weight_percent, (int, float)) else 0
-    
-    @property
-    def remaining_filter_days(self) -> int:
-        """Get the number of days remaining for the filter replacement."""
-        return self._data.get("realInfo", {}).get("remainingReplacementDays", 0)
-    
-    @property
-    def remaining_cleaning_days(self) -> int:
-        """Get the number of days remaining for machine cleaning."""
-        return self._data.get("realInfo", {}).get("remainingCleaningDays", 0)
-    
-    @property
-    def light_switch(self) -> bool:
-        """Check if the light is enabled."""
-        return bool(self._data.get("realInfo", {}).get("lightSwitch", False))
-    
-    @property
-    def sound_switch(self) -> bool:
-        """Check if the sound is enabled."""
-        return self._data.get("realInfo", {}).get("soundSwitch", False)
-    
-    async def set_light_switch(self, value: bool):
-        """Enable or disable the light."""
-        await self.api.set_light_switch(self.serial, value)
-        await self.refresh()
-    
-    async def set_sound_switch(self, value: bool):
-        """Enable or disable the sound."""
-        await self.api.set_sound_switch(self.serial, value)
-        await self.refresh()
+    # ── properties unique to Cordless fountain ────────────────────
 
     @property
     def detection_sensitivity(self) -> str:
         """Get the detection sensitivity."""
         return self._data.get("dataRealInfo", {}).get("radarSensingLevel", "unknown")
-
-    @property
-    def water_interval(self) -> float:
-        water_interval = self._data.get("realInfo", {}).get("useWaterInterval")
-        return water_interval if isinstance(water_interval, (int, float)) else 0
-    @property
-    def water_dispensing_duration(self) -> float:
-        duration = self._data.get("realInfo", {}).get("useWaterDuration")
-        return duration if isinstance(duration, (int, float)) else 0
 
     @property
     def water_state(self) -> bool:
@@ -132,19 +49,16 @@ class Dockstream2SmartCordlessFountain(Device):
         """Get current water dispensing mode."""
         real = self._data.get("dataRealInfo", {}) or {}
 
-        # raw values as received
         stop_raw = real.get("waterStopSwitch")
         mode_raw = real.get("useWaterType")
         radar = real.get("radarSensingLevel")
 
-        # coerce to the types we expect
         stop = bool(stop_raw)
         try:
             mode = int(mode_raw) if mode_raw is not None else None
         except (TypeError, ValueError):
             mode = None
 
-        # Decide label
         if stop:
             label = "Off"
         elif mode == 0:
@@ -155,14 +69,14 @@ class Dockstream2SmartCordlessFountain(Device):
             elif radar == "FarTrigger":
                 label = "Sensor-Activated (Far)"
             else:
-                label = "Unknown"  # waiting for radar to land
+                label = "Unknown"
         else:
             label = "Unknown"
 
         return label
 
     # Not currently supported by the device, API accepts, but device doesnt apply. hoping for future firmware update.
-    
+
     # @property
     # def water_sensing_delay(self) -> float:
     #     return self._data.get("dataRealInfo", {}).get("sensingWaterDuration", 0)
@@ -186,58 +100,10 @@ class Dockstream2SmartCordlessFountain(Device):
         _LOGGER.debug(f"Setting water low threshold to {value} for {self.serial}")
         try:
             await self.api.set_water_low_threshold(self.serial, value)
-            await self.refresh()  # Refresh the state after the action
+            await self.refresh()
         except aiohttp.ClientError as err:
             _LOGGER.error(f"Failed to set water low threshold to {value} for {self.serial}: {err}")
             raise PetLibroAPIError(f"Error setting water low threshold: {err}")
-
-    @property
-    def cleaning_cycle(self) -> float:
-        cleaning_cycle = self._data.get("realInfo", {}).get("machineCleaningFrequency")
-        return cleaning_cycle if isinstance(cleaning_cycle, (int, float)) else 0
-
-    async def set_cleaning_cycle(self, value: float) -> None:
-        _LOGGER.debug(f"Setting machine cleaning cycle to {value} for {self.serial}")
-        try:
-            key = "MACHINE_CLEANING"
-            await self.api.set_filter_cycle(self.serial, value, key)
-            await self.refresh()  # Refresh the state after the action
-        except aiohttp.ClientError as err:
-            _LOGGER.error(f"Failed to set cleaning cycle using {key} for {self.serial}: {err}")
-            raise PetLibroAPIError(f"Error setting cleaning cycle using {key}: {err}")
-
-    @property
-    def filter_cycle(self) -> float:
-        filter_cycle = self._data.get("realInfo", {}).get("filterReplacementFrequency")
-        return filter_cycle if isinstance(filter_cycle, (int, float)) else 0
-
-    async def set_filter_cycle(self, value: float) -> None:
-        _LOGGER.debug(f"Setting filter cycle to {value} for {self.serial}")
-        try:
-            key = "FILTER_ELEMENT"
-            await self.api.set_filter_cycle(self.serial, value, key)
-            await self.refresh()  # Refresh the state after the action
-        except aiohttp.ClientError as err:
-            _LOGGER.error(f"Failed to set filter cycle using {key} for {self.serial}: {err}")
-            raise PetLibroAPIError(f"Error setting filter cycle using {key}: {err}")
-
-    async def set_cleaning_reset(self) -> None:
-        _LOGGER.debug(f"Triggering machine cleaning reset for {self.serial}")
-        try:
-            await self.api.set_cleaning_reset(self.serial)
-            await self.refresh()  # Refresh the state after the action
-        except aiohttp.ClientError as err:
-            _LOGGER.error(f"Failed to trigger machine cleaning reset for {self.serial}: {err}")
-            raise PetLibroAPIError(f"Error triggering machine cleaning reset: {err}")
-
-    async def set_filter_reset(self) -> None:
-        _LOGGER.debug(f"Triggering filter reset for {self.serial}")
-        try:
-            await self.api.set_filter_reset(self.serial)
-            await self.refresh()  # Refresh the state after the action
-        except aiohttp.ClientError as err:
-            _LOGGER.error(f"Failed to trigger filter reset for {self.serial}: {err}")
-            raise PetLibroAPIError(f"Error triggering filter reset: {err}")
 
     @property
     def battery_state(self) -> str:
@@ -258,8 +124,7 @@ class Dockstream2SmartCordlessFountain(Device):
     @property
     def power_state(self) -> int:
         api_value = self._data.get("dataRealInfo", {}).get("powerType", 0)
-        
-        # Direct mapping inside the property
+
         if api_value == 2:
             return False
         elif api_value == 3:
@@ -270,104 +135,3 @@ class Dockstream2SmartCordlessFountain(Device):
     @property
     def electric_quantity(self) -> int:
         return self._data.get("realInfo", {}).get("electricQuantity", 0)
-
-    @property
-    def today_drinking_amount(self) -> float:
-        """Get the total milliliters of water used today."""
-        amount = self._data.get("getDrinkWater", {}).get("todayTotalMl")
-        return amount if isinstance(amount, (int, float)) else 0
-    
-    @property
-    def today_drinking_count(self) -> int:
-        """Get the total count of times drank today."""
-        drinking_count = self._data.get("getDrinkWater", {}).get("todayTotalTimes")
-        return drinking_count if isinstance(drinking_count, int) else 0
-
-    @property
-    def today_drinking_time(self) -> int:
-        """Get the total time spent drinking today."""
-        drinking_time = self._data.get("getDrinkWater", {}).get("petEatingTime")
-        return drinking_time if isinstance(drinking_time, int) else 0
-
-    @property
-    def today_avg_time(self) -> int:
-        """Get the average time spent drinking in a session today."""
-        avg_time = self._data.get("getDrinkWater", {}).get("avgDrinkDuration")
-        return avg_time if isinstance(avg_time, int) else 0
-
-    @property
-    def yesterday_drinking_amount(self) -> float:
-        """Get the total milliliters of water used yesterday."""
-        amount = self._data.get("getDrinkWater", {}).get("yesterdayTotalMl")
-        return amount if isinstance(amount, (int, float)) else 0
-    
-    @property
-    def yesterday_drinking_count(self) -> int:
-        """Get the total count of times drank yesterday."""
-        drinking_count = self._data.get("getDrinkWater", {}).get("yesterdayTotalTimes")
-        return drinking_count if isinstance(drinking_count, int) else 0
-
-    @property
-    def filter_replacement_frequency(self) -> int:
-        """Get the filter replacement frequency."""
-        frequency = self._data.get("realInfo", {}).get("filterReplacementFrequency")
-        return frequency if isinstance(frequency, int) else 0
-    
-    @property
-    def machine_cleaning_frequency(self) -> int:
-        """Get the machine cleaning frequency."""
-        frequency = self._data.get("realInfo", {}).get("machineCleaningFrequency")
-        return frequency if isinstance(frequency, int) else 0
-
-    # Method for indicator turn on
-    async def set_light_on(self) -> None:
-        _LOGGER.debug(f"Turning on the indicator for {self.serial}")
-        try:
-            await self.api.set_light_on(self.serial)
-            await self.refresh()  # Refresh the state after the action
-        except aiohttp.ClientError as err:
-            _LOGGER.error(f"Failed to turn on the indicator for {self.serial}: {err}")
-            raise PetLibroAPIError(f"Error turning on the indicator: {err}")
-
-    # Method for indicator turn off
-    async def set_light_off(self) -> None:
-        _LOGGER.debug(f"Turning off the indicator for {self.serial}")
-        try:
-            await self.api.set_light_off(self.serial)
-            await self.refresh()  # Refresh the state after the action
-        except aiohttp.ClientError as err:
-            _LOGGER.error(f"Failed to turn off the indicator for {self.serial}: {err}")
-            raise PetLibroAPIError(f"Error turning off the indicator: {err}")
-
-    @property
-    def update_available(self) -> bool:
-        """Return True if an update is available, False otherwise."""
-        return bool(self._data.get("getUpgrade", {}).get("jobItemId"))
-    
-    @property
-    def update_release_notes(self) -> str | None:
-        """Return release notes if available, else None."""
-        upgrade_data = self._data.get("getUpgrade")
-        return upgrade_data.get("upgradeDesc") if upgrade_data else None
-    
-    @property
-    def update_version(self) -> str | None:
-        """Return target version if available, else None."""
-        upgrade_data = self._data.get("getUpgrade")
-        return upgrade_data.get("targetVersion") if upgrade_data else None
-    
-    @property
-    def update_name(self) -> str | None:
-        """Return update job name if available, else None."""
-        upgrade_data = self._data.get("getUpgrade")
-        return upgrade_data.get("jobName") if upgrade_data else None
-    
-    @property
-    def update_progress(self) -> float:
-        """Return update progress as a float, or 0 if not updating."""
-        upgrade_data = self._data.get("getUpgrade")
-        if not upgrade_data:
-            return 0.0
-
-        progress = upgrade_data.get("progress")
-        return float(progress) if progress is not None else 0.0
