+++ custom_components/petlibro/manifest.json
@@ -12,5 +12,5 @@
   "issue_tracker": "https://github.com/jjjonesjr33/petlibro/issues",
   "loggers": ["petlibro"],
   "requirements": [],
-  "version": "1.2.30.8"
+  "version": "1.3.0.0"
 }
